chore: updates github provider tests
